requirements.txt
# Night-Walk MVP - Requirements
# Python 3.10+

# Flask Core
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1

# Database
SQLAlchemy>=2.0.30
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL (production)

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# External APIs
twilio==8.10.0
stripe==7.8.0

# Scheduler
APScheduler==3.10.4

# Utilities
gunicorn==21.2.0
Pillow>=10.0.0

# Security
Flask-Limiter==3.5.0
python-magic-bin==0.4.14;sys_platform=='win32'
python-magic==0.4.27;sys_platform!='win32'

# PDF Generation (Render compatible)
xhtml2pdf==0.2.11

# Email via SendGrid
sendgrid==6.10.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
